# find-my-mua/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Find My MUA API",
        "description": "Dokumentasi lengkap API untuk aplikasi Find My MUA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/mua/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Lihat profil MUA (private)",
                "operationId": "5d6c4c4a9bcb4ee1288d7c566db3d2fe",
                "responses": {
                    "200": {
                        "description": "Profil MUA"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profil MUA",
                "operationId": "26bac6f6c19ba7f0527ff5c0251c0ad7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bio": {
                                        "type": "string",
                                        "example": "Certified bridal MUA"
                                    },
                                    "makeup_styles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "makeup_specializations": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mua/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Lihat profil publik MUA (portofolio, layanan, review)",
                "operationId": "b8a4625af8a6d80324d6675f5b9e9abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil publik dikembalikan"
                    }
                }
            }
        },
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Ambil notifikasi pengguna",
                "operationId": "77ed4ddee94926a5e8732c891c460264",
                "responses": {
                    "200": {
                        "description": "List notifikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/read": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Tandai semua notifikasi sebagai sudah dibaca",
                "operationId": "4b3593831fc028c63931fc686d61d9d4",
                "responses": {
                    "200": {
                        "description": "Berhasil ditandai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}